fix compatibility with debezium 1.2.0.Final
Fix materializedstate associated with the event. Previously two event was emitted in kafka: a creational one an updated one (when materializedstate was added to the event)
Now only one event is produced having the event payload, metadata and materializedstate.

Rework and add test (notably one checking event is well produce in kafka topic)